Start refactoring renderer files
