+++ src/engine/renderer/material.ts
@@ -12,7 +12,7 @@ import {
 } from "three";
 
 import { getModule } from "../module/module.common";
-import { RendererModule, RenderContext } from "../renderer/renderer.render";
+import { RendererModule, RenderContext } from "./renderer.render";
 import { RenderMaterial, RenderMeshPrimitive } from "../resource/resource.render";
 import { MeshPrimitiveAttributeIndex, MeshPrimitiveMode } from "../resource/schema";
 import { MatrixMaterial } from "./MatrixMaterial";
