+++ src/engine/renderer/patchShaderChunks.ts
