+++ src/engine/renderer/systems/RendererIncomingTripleBufferSystem.ts
@@ -1,5 +1,5 @@
-import { swapReadBufferFlags } from "../allocator/TripleBuffer";
-import { RenderContext } from "./renderer.render";
+import { swapReadBufferFlags } from "../../allocator/TripleBuffer";
+import { RenderContext } from "../renderer.render";
 
 export function RendererIncomingTripleBufferSystem(ctx: RenderContext) {
   const bufferSwapped = swapReadBufferFlags(ctx.gameToRenderTripleBufferFlags);
