+++ src/engine/renderer/systems/RendererOutgoingTripleBufferSystem.ts
@@ -1,5 +1,5 @@
-import { swapWriteBufferFlags } from "../allocator/TripleBuffer";
-import { RenderContext } from "./renderer.render";
+import { swapWriteBufferFlags } from "../../allocator/TripleBuffer";
+import { RenderContext } from "../renderer.render";
 
 export function RendererOutgoingTripleBufferSystem(ctx: RenderContext) {
   swapWriteBufferFlags(ctx.renderToGameTripleBufferFlags);
